+++ src/components/Header.jsx
@@ -2,44 +2,25 @@ import { useState } from "react";
 import MobileNav from "./header/mobile/MobileNav";
 import DesktopNav from "./header/desktop/DesktopNav";
 import NavLogo from "./header/NavLogo";
-import useAuth from "../hooks/useAuth";
-import { useLocation } from "react-router";
-import { Link } from "react-router-dom";
+import LoggedInUser from "./header/LoggedInUser";
 
 export default function Header() {
   const [nav, setNav] = useState(true);
-  const { auth } = useAuth();
-  const location = useLocation();
 
   const handleNav = () => {
     setNav(!nav);
   };
 
   return (
-    <>
+    <div className="text-[16px] font-sans">
       <div className="max-w-[1240px] mx-auto px-4">
-        {auth?.accessToken && location.pathname !== "/" ? (
-          <p className="text-right h-8 text-[14px] ">
-            <span className="hidden md:block">
-              Bruger:{" "}
-              {auth?.user_roles.includes("admin") ? (
-                <span className="font-bold">{auth.user_email}</span>
-              ) : (
-                <Link to="/profile">
-                  <span className="font-bold">{auth.user_email}</span>
-                </Link>
-              )}
-            </span>
-          </p>
-        ) : (
-          <p className="text-right h-8 text-[14px]"></p>
-        )}
+        <LoggedInUser />
       </div>
       <div className="flex justify-between items-center h-18 max-w-[1240px] mx-auto px-4 pb-4 text-black uppercase font-medium">
         <NavLogo />
         <DesktopNav />
         <MobileNav handleNav={handleNav} nav={nav} />
       </div>
-    </>
+    </div>
   );
 }
