+++ src/components/header/desktop/LandingPageNavItems.jsx
@@ -1,0 +1,12 @@
+import DesktopNavItem from "./DesktopNavItem";
+
+function LandingPageNavItems() {
+  return (
+    <>
+      <DesktopNavItem path="/shop">Dagligvarer</DesktopNavItem>
+      <DesktopNavItem path="/about">Om os</DesktopNavItem>
+    </>
+  );
+}
+
+export default LandingPageNavItems;
